requirements.txt
streamlit
langchain
langchain-openai
langchain-community
chromadb
pandas
python-dotenv 